Web Client

+++ Spring-11-Rest-Jackson/src/main/java/com/cybertek/controller/WebFluxController.java
@@ -104,11 +104,15 @@ public class WebFluxController {
                 .bodyToMono(Genre.class);
     }
 
-    @DeleteMapping("/delete/id")
-    public Mono<Void> deleteWebClient(@PathVariable("id") Long id){
+    @DeleteMapping("/delete/{id}")
+    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) throws Exception{
+        Integer countGenres = genreRepository.countGenresNativeQuery(id);
+        if(countGenres > 0){
+          throw new Exception("Genre can't be deleted, is linked by a movie");
+        }
         return webClient
                 .delete()
-                .uri("delete-genre/{id}",id)
+                .uri("/delete-genre/{id}",id)
                 .retrieve()
                 .bodyToMono(Void.class);
     }
